refactor(storyboard): 把分镜方案的提示词构造拆出 storyboardPlanPrompts
storyboardPlan.ts 在 origin/main（#494 合入后）已是 818 行，check:filesize 在 main 上就是红的，
本分支并未改这个文件。按 R9/R12 把纯函数那一段（锚身份描述 / 定妆卡版面提示词 / 镜头与首帧
最终 prompt 合成 / 视觉参考边顺序 / isVisualAnchor）原样搬进同目录新模块，行为逐字不变：

- 818 → 696 行，新模块 144 行，两边都在上限内
- 类型仍住 storyboardPlan.ts，新模块用 import type 引用（编译期擦除，不构成运行时环；check:boundaries 无新增）
- buildAnchorSheetPrompt / isVisualAnchor 原样再导出，调用方 import 路径一个字不改
- storyboardPlan.test.ts 443 条原样绿
